requirements.txt
Pillow>=8.0.0
argparse 